requirements.txt
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
python-dotenv==0.14.0
